refactor(vite-web-widgets): split shared vite config into domain modules
extract config creation/resolution into config/*
extract mpk/editor build steps into build/*
extract package/json helpers into helpers/*
centralize shared types in types.ts
keep both direct CLI and programmatic createWidgetViteConfig compatibility
update README to document module layout and usage modes
