+++ packages/shared/vite-web-widgets/vite.config.ts
@@ -1,327 +1,11 @@
-import Archiver from "archiver";
-import { copyFileSync, createWriteStream, existsSync, mkdirSync, readFileSync, rmSync } from "fs";
-import { cp } from "fs/promises";
-import { join, resolve } from "path";
-import { build as viteBuild, defineConfig, type UserConfig } from "vite";
+import { defineConfig, type ConfigEnv } from "vite";
+import { createConfig } from "./config/create";
+import type { WidgetViteConfigOptions } from "./types";
 
-type EditorBuild = {
-    entry: string;
-    outputFile: string;
-    externals: Array<string | RegExp>;
-    format?: "cjs" | "es";
-};
-
-type RuntimeOutput = {
-    format: "cjs" | "es" | "amd";
-    entryFileName: string;
-};
-
-type FileCopy = {
-    src: string;
-    dest: string;
-};
-
-type WidgetPackageJson = {
-    name: string;
-    version: string;
-    packagePath: string;
-    mxpackage?: {
-        mpkName?: string;
-    };
-};
-
-type WidgetViteConfigOptions = {
-    widgetName: string;
-    runtimeDirectoryName?: string;
-};
-
-function getResolveAlias(): { find: RegExp; replacement: string }[] {
-    return [
-        {
-            find: /^~(.+)/,
-            replacement: "$1"
-        },
-        {
-            find: /^src\//,
-            replacement: `${resolve(process.cwd(), "src")}/`
-        }
-    ];
+export function createWidgetViteConfig(options: WidgetViteConfigOptions = {}) {
+    return defineConfig((env: ConfigEnv) => createConfig(options, env));
 }
 
-async function copyDir(src: string, dest: string): Promise<void> {
-    mkdirSync(dest, { recursive: true });
-    if (existsSync(src)) {
-        await cp(src, dest, { recursive: true, force: true });
-    }
-}
-
-async function buildEditorArtifacts(editorBuilds: EditorBuild[]): Promise<void> {
-    const editorOutDir = "dist/tmp/widgets";
-    const alias = getResolveAlias();
-
-    for (const editorBuild of editorBuilds) {
-        await viteBuild({
-            configFile: false,
-            resolve: {
-                alias
-            },
-            build: {
-                target: "es2019",
-                minify: "esbuild",
-                emptyOutDir: false,
-                outDir: editorOutDir,
-                lib: {
-                    entry: editorBuild.entry,
-                    formats: [editorBuild.format ?? "cjs"],
-                    fileName: () => editorBuild.outputFile
-                },
-                rollupOptions: {
-                    external: editorBuild.externals,
-                    output: {
-                        format: editorBuild.format ?? "cjs",
-                        entryFileNames: editorBuild.outputFile,
-                        inlineDynamicImports: true
-                    }
-                }
-            }
-        });
-    }
-}
-
-function readWidgetPackageJson(): WidgetPackageJson {
-    const packageJsonPath = resolve(process.cwd(), "package.json");
-    const packageJsonText = readFileSync(packageJsonPath, "utf-8");
-    return JSON.parse(packageJsonText) as WidgetPackageJson;
-}
-
-function toPackagePathDir(packagePath: string): string {
-    return packagePath.replace(/\./g, "/");
-}
-
-function inferPrimaryRuntimeFormat(): "cjs" | "amd" {
-    if (process.env.VITE_RUNTIME_FORMAT === "cjs") {
-        return "cjs";
-    }
-
-    return "amd";
-}
-
-function inferMetadataFiles(widgetName: string): FileCopy[] {
-    return [
-        { src: `src/${widgetName}.xml`, dest: `${widgetName}.xml` },
-        { src: `src/${widgetName}.icon.png`, dest: `${widgetName}.icon.png` },
-        { src: `src/${widgetName}.icon.dark.png`, dest: `${widgetName}.icon.dark.png` },
-        { src: `src/${widgetName}.tile.png`, dest: `${widgetName}.tile.png` },
-        { src: `src/${widgetName}.tile.dark.png`, dest: `${widgetName}.tile.dark.png` },
-        { src: "../../../LICENSE", dest: "License.txt" },
-        { src: "src/package.xml", dest: "package.xml" }
-    ];
-}
-
-function inferRequiredArtifacts(
-    widgetName: string,
-    packagePath: string,
-    runtimeDirectoryName: string,
-    editorBuilds: EditorBuild[]
-): string[] {
-    const packagePathDir = toPackagePathDir(packagePath);
-    const widgetDir = runtimeDirectoryName;
-
-    const editorArtifacts = editorBuilds.map(editorBuild => editorBuild.outputFile);
-
-    return [
-        ...editorArtifacts,
-        `${packagePathDir}/${widgetDir}/${widgetName}.js`,
-        `${packagePathDir}/${widgetDir}/${widgetName}.mjs`
-    ];
-}
-
-function inferRuntimeOutDir(packagePath: string, runtimeDirectoryName: string): string {
-    const packagePathDir = toPackagePathDir(packagePath);
-    return `dist/tmp/widgets/${packagePathDir}/${runtimeDirectoryName}`;
-}
-
-function inferEditorBuilds(widgetName: string): EditorBuild[] {
-    const editorBuilds: EditorBuild[] = [];
-
-    const editorPreviewEntry = `src/${widgetName}.editorPreview.tsx`;
-    if (existsSync(editorPreviewEntry)) {
-        editorBuilds.push({
-            entry: editorPreviewEntry,
-            outputFile: `${widgetName}.editorPreview.js`,
-            externals: [/^mendix($|\/)/, /^react$/, /^react-dom$/]
-        });
-    }
-
-    const editorConfigEntry = `src/${widgetName}.editorConfig.ts`;
-    if (existsSync(editorConfigEntry)) {
-        editorBuilds.push({
-            entry: editorConfigEntry,
-            outputFile: `${widgetName}.editorConfig.js`,
-            externals: [/^mendix($|\/)/, /^react$/, /^react-dom$/]
-        });
-    }
-
-    return editorBuilds;
-}
-
-function inferRemoveBeforeCopy(packageName: string): string[] {
-    const widgetPackageName = packageName.split("/").pop();
-    return widgetPackageName ? [`${widgetPackageName}.css`] : [];
-}
-
-type ResolvedConfig = {
-    widgetName: string;
-    widgetVersion: string;
-    mpkName: string;
-    runtimeEntry: string;
-    runtimeOutDir: string;
-    runtimeOutputs: RuntimeOutput[];
-    runtimeExternals: Array<string | RegExp>;
-    metadataFiles: FileCopy[];
-    editorBuilds: EditorBuild[];
-    requiredArtifacts: string[];
-    removeBeforeCopy: string[];
-    define: Record<string, string>;
-};
-
-function resolveConfig(options: WidgetViteConfigOptions): ResolvedConfig {
-    const widgetPackageJson = readWidgetPackageJson();
-    const primaryRuntimeFormat = inferPrimaryRuntimeFormat();
-    const editorBuilds = inferEditorBuilds(options.widgetName);
-    const runtimeDirectoryName = options.runtimeDirectoryName ?? options.widgetName.toLowerCase();
-
-    return {
-        widgetName: options.widgetName,
-        widgetVersion: widgetPackageJson.version,
-        mpkName: widgetPackageJson.mxpackage?.mpkName ?? `${options.widgetName}.mpk`,
-        runtimeEntry: `src/${options.widgetName}.tsx`,
-        runtimeOutDir: inferRuntimeOutDir(widgetPackageJson.packagePath, runtimeDirectoryName),
-        runtimeOutputs: [
-            {
-                format: primaryRuntimeFormat,
-                entryFileName: `${options.widgetName}.js`
-            },
-            {
-                format: "es",
-                entryFileName: `${options.widgetName}.mjs`
-            }
-        ],
-        runtimeExternals: ["react", "react-dom", "@mendix/widget-plugin-component-kit", "big.js", /^mendix($|\/)/],
-        metadataFiles: inferMetadataFiles(options.widgetName),
-        editorBuilds,
-        requiredArtifacts: inferRequiredArtifacts(
-            options.widgetName,
-            widgetPackageJson.packagePath,
-            runtimeDirectoryName,
-            editorBuilds
-        ),
-        removeBeforeCopy: inferRemoveBeforeCopy(widgetPackageJson.name),
-        define: {
-            "process.env.NODE_ENV": JSON.stringify("production")
-        }
-    };
-}
-
-async function createMPK(options: ResolvedConfig): Promise<void> {
-    const distPath = resolve(process.cwd(), "dist");
-    const tmpWidgetsPath = join(distPath, "tmp", "widgets");
-    const stagingDir = join(distPath, "widgets");
-    const outputDir = join(process.cwd(), "dist", options.widgetVersion);
-    const mpkPath = join(outputDir, options.mpkName);
-
-    if (existsSync(stagingDir)) {
-        rmSync(stagingDir, { recursive: true });
-    }
-    mkdirSync(stagingDir, { recursive: true });
-    mkdirSync(outputDir, { recursive: true });
-
-    for (const file of options.metadataFiles) {
-        const srcPath = resolve(process.cwd(), file.src);
-        const destPath = join(stagingDir, file.dest);
-        mkdirSync(join(stagingDir, file.dest.split("/").slice(0, -1).join("/")), { recursive: true });
-        if (existsSync(srcPath)) {
-            copyFileSync(srcPath, destPath);
-        }
-    }
-
-    for (const requiredArtifact of options.requiredArtifacts ?? []) {
-        const requiredPath = join(tmpWidgetsPath, requiredArtifact);
-        if (!existsSync(requiredPath)) {
-            throw new Error(`Missing compiled artifact: ${requiredPath}`);
-        }
-    }
-
-    for (const removePath of options.removeBeforeCopy ?? []) {
-        const absolutePath = join(tmpWidgetsPath, removePath);
-        if (existsSync(absolutePath)) {
-            rmSync(absolutePath);
-        }
-    }
-
-    if (existsSync(tmpWidgetsPath)) {
-        await copyDir(tmpWidgetsPath, stagingDir);
-    }
-
-    await new Promise<void>((resolvePromise, reject) => {
-        const output = createWriteStream(mpkPath);
-        const archive = (Archiver as unknown as (format: string, options: { zlib: { level: number } }) => any)("zip", {
-            zlib: { level: 9 }
-        });
-
-        output.on("close", () => {
-            console.log(`✓ Created ${mpkPath} (${archive.pointer()} bytes)`);
-            resolvePromise();
-        });
-
-        archive.on("error", reject);
-        archive.pipe(output);
-        archive.directory(stagingDir, false);
-        archive.finalize();
-    });
-}
-
-export default function createWidgetViteConfig(options: WidgetViteConfigOptions): UserConfig {
-    const resolvedConfig = resolveConfig(options);
-    const alias = getResolveAlias();
-
-    return defineConfig({
-        define: resolvedConfig.define,
-        resolve: {
-            alias
-        },
-        build: {
-            target: "es2019",
-            minify: "esbuild",
-            lib: {
-                entry: resolvedConfig.runtimeEntry,
-                name: resolvedConfig.widgetName
-            },
-            outDir: resolvedConfig.runtimeOutDir,
-            rollupOptions: {
-                output: resolvedConfig.runtimeOutputs.map(runtimeOutput => ({
-                    format: runtimeOutput.format,
-                    entryFileNames: runtimeOutput.entryFileName,
-                    inlineDynamicImports: true
-                })),
-                external: resolvedConfig.runtimeExternals
-            }
-        },
-        plugins: [
-            {
-                name: "vite-plugin-mpk-builder",
-                apply: "build",
-                enforce: "post",
-                async closeBundle() {
-                    if (resolvedConfig.editorBuilds.length > 0) {
-                        console.log("Building editor artifacts...");
-                        await buildEditorArtifacts(resolvedConfig.editorBuilds);
-                    }
-
-                    console.log("Building MPK...");
-                    await createMPK(resolvedConfig);
-                }
-            }
-        ]
-    });
-}
+// Default export supports direct CLI usage:
+// `vite build --config ../../shared/vite-web-widgets/vite.config.ts`
+export default defineConfig((env: ConfigEnv) => createConfig({}, env));
